feat: re-add /join, both /join and /login use the same SignIn component
